+++ src/backend/InvenTree/InvenTree/api_version.py
@@ -1,10 +1,14 @@
 """InvenTree API version information."""
 
 # InvenTree API version
-INVENTREE_API_VERSION = 455
+INVENTREE_API_VERSION = 456
 """Increment this API version number whenever there is a significant change to the API that any clients need to know about."""
 
 INVENTREE_API_TEXT = """
+
+v456 -> 2026-02-20 : https://github.com/inventree/InvenTree/pull/11303
+    - Adds "primary" field to the SupplierPart API
+    - Removes "default_supplier" field from the Part API
 
 v455 -> 2026-02-19 : https://github.com/inventree/InvenTree/pull/11383
     - Adds "exists_for_model_id" filter to ParameterTemplate API endpoint